Contract acts now inherit from a base class
